refactor(core+backend/X11): 将修改窗口几何属性合并到 CONFIGURE_WINDOW 中
将 MOVE/RESIZE/BORDER_WIDTH 三个效果合并进 CONFIGURE_WINDOW 中，通过
changed_fields 掩码提取有效数据

+++ src/backend/x11/internal.h
@@ -60,7 +60,7 @@ typedef struct atoms_t {
 
 typedef struct window_configure_t {
   xcb_window_t window;
-  uint16_t mask;
+  uint32_t mask;
   xcb_configure_window_value_list_t value;
 } window_configure_t;
